+++ JuliaBUGS/src/model/Model.jl
@@ -20,9 +20,14 @@ include("evaluation.jl")
 include("abstractppl.jl")
 include("logdensityproblems.jl")
 
-export parameters, variables, initialize!, getparams, settrans, set_evaluation_mode
-export regenerate_log_density_function, set_observed_values!
-export evaluate_with_rng!!, evaluate_with_env!!, evaluate_with_values!!
-export BUGSModelWithGradient, _logdensity_switched
+# Public user-facing API
+export parameters, variables, initialize!, getparams, settrans
+export set_evaluation_mode, set_observed_values!
+
+# Evaluation mode types
+export UseGraph, UseGeneratedLogDensityFunction
+
+# Gradient wrapper
+export BUGSModelWithGradient
 
 end # Model